pyproject.toml
[project]
name = "repl-nix-workspace"
version = "2.17"
description = "Test create project with 3rd party rules of creating - I haven't seen bare-agent before, so I don't know the difference- if this set of rules is better than au natural, I won't immediately know. My first project ever."
requires-python = ">=3.11"
dependencies = [
    "apify-client>=1.12.0",
    "black>=25.1.0",
    "bleach>=6.2.0",
    "email-validator>=2.2.0",
    "flake8>=7.3.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "google-api-python-client>=2.176.0",
    "google-auth-oauthlib>=1.2.2",
    "google-genai>=1.26.0",
    "google-generativeai>=0.8.5",
    "gunicorn>=23.0.0",
    "isort>=6.0.1",
    "numpy>=2.3.1",
    "openai>=1.93.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pytest>=8.4.2",
    "python-docx>=1.2.0",
    "replit-object-storage>=1.0.2",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.41",
    "tenacity>=8.5.0",
    "trafilatura>=2.0.0",
    "vulture>=2.14",
    "werkzeug>=3.1.3",
]
